Script for Extracting Pore residues
Wrote a TCL script to extract the residues facing into the pore. In order to make it work, I first need to identify the axis of the pore before setting the trajectory of a sodium ion along it. Thereonafter, it extracts residues within certain distance of the SOD ion and makes a list of unique residues it came in contact with.

Note: Adjust the distance parameter based on how large you want your pore to be considered.